changing the js files structure
