+++ js/util.js
@@ -1,0 +1,7 @@
+// return number from 'start' to 'end'
+const getRadomNumber = (start, end) =>
+  Math.round(Math.random() * (end - start) + start);
+// return true if string's length less or equals when 'maxLen'
+const checkStringLen = (str, maxLen) => str.length <= maxLen;
+
+export { getRadomNumber, checkStringLen };
